pyproject.toml
[tool.poetry]
name = "useful-habits"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
setuptools = "^72.1.0"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.6.0"
eventlet = "^0.36.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
